Split tests that require ledger support
